autenticação

+++ miniEcommerceSpring/src/main/java/com/app/vdc/demo/repository/CarrinhoRepository.java
@@ -3,7 +3,7 @@ package com.app.vdc.demo.repository;
 import org.springframework.data.repository.CrudRepository;
 import com.app.vdc.demo.Model.Carrinho;
 
-public interface CarrinhoRepository extends CrudRepository<Carrinho, Integer>{
+public interface CarrinhoRepository extends CrudRepository<Carrinho, Long>{
 
 	
 }
