+++ miniEcommerceSpring/src/main/java/com/app/vdc/demo/repository/PedidoRepository.java
@@ -3,7 +3,7 @@ package com.app.vdc.demo.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.app.vdc.demo.Model.Pedido;
 
-public interface PedidoRepository extends JpaRepository<Pedido,Integer> {
+public interface PedidoRepository extends JpaRepository<Pedido, Long> {
       
 
 }
